이미지로 bg 설정 및 페이지별 layout 분리
