+++ src/App.tsx
@@ -5,7 +5,8 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import '@/font.css';
 import '@/index.css';
-import CommonLayout from '@/layout/CommonLayout';
+import CommonLayout from '@/layouts/CommonLayout';
+import HomeLayout from '@/layouts/HomeLayout';
 import Create from '@/pages/Create';
 import Home from '@/pages/Home';
 import Room from '@/pages/Room';
@@ -18,8 +19,10 @@ function App() {
     <QueryClientProvider client={queryClient}>
       <BrowserRouter>
         <Routes>
-          <Route element={<CommonLayout />}>
+          <Route element={<HomeLayout />}>
             <Route path="/" element={<Home />} />
+          </Route>
+          <Route element={<CommonLayout />}>
             <Route path="/create" element={<Create />} />
             <Route path="/survey" element={<Survey />} />
             <Route path="/:roomId" element={<Room />} />
